*: reorg oracle expire test
